Add files via upload
This Python script includes the pipeline I designed to gather and process the neccessary datasets related to crowd analysis and event management. A wide range of data types and file types had to be processed, including CSV files, images, videos, and MATLAB .mat files. This pipeline can work with any of those types of data sources or file types.